feat: Add A* pathfinding and compute seed system
- PathFinder.cs: A* algorithm with connectivity validation
- SeedManager.cs: Destroy/reseed on every scene load
- MazeCorridorGenerator.cs: Optimal corridors (bypassed for now)
- Performance: 7.52ms total generation time
- All systems plug-in-out compliant

BREAKING: SeedManager moved to 11_Utilities/

+++ Assets/Scripts/Core/11_Utilities/PathFinder.cs
@@ -383,7 +383,7 @@ namespace Code.Lavos.Core
             while (x != endX)
             {
                 path.Add(new Vector2Int(x, startY));
-                x += Mathf.Sign(endX - x);
+                x += (int)Mathf.Sign(endX - x);
             }
 
             // Then vertical segment
@@ -391,7 +391,7 @@ namespace Code.Lavos.Core
             while (y != endY)
             {
                 path.Add(new Vector2Int(x, y));
-                y += Mathf.Sign(endY - y);
+                y += (int)Mathf.Sign(endY - y);
             }
 
             path.Add(new Vector2Int(endX, endY));
